+++ tests/end_to_end/loop_with_if/irix-g-out.c
@@ -8,11 +8,11 @@ s32 test(s32 arg0)
 loop_1:
         if (sp4 == 5)
         {
-            sp4 = (s32) (sp4 * 2);
+            sp4 = sp4 * 2;
         }
         else
         {
-            sp4 = (s32) (sp4 + 4);
+            sp4 = sp4 + 4;
         }
         if (sp4 < arg0)
         {
