+++ tests/end_to_end/sqrt/irix-g-out.c
@@ -6,11 +6,11 @@ f32 test(f32 arg0)
     f32 sp0;
 
     sp0 = 1.5f;
-    sp8 = (f32) (arg0 * 0.5f);
+    sp8 = arg0 * 0.5f;
     sp4 = arg0;
-    spC = (s32) sp4;
-    spC = (s32) (0x5F3759DF - (spC >> 1));
-    sp4 = (f32) spC;
-    sp4 = (f32) ((sp0 - ((sp8 * sp4) * sp4)) * sp4);
+    spC = sp4;
+    spC = 0x5F3759DF - (spC >> 1);
+    sp4 = spC;
+    sp4 = (sp0 - ((sp8 * sp4) * sp4)) * sp4;
     return sp4;
 }
