+++ tests/end_to_end/struct/irix-g-out.c
@@ -2,7 +2,7 @@ void *test(void *arg0)
 {
     s32 sp4;
 
-    sp4 = (s32) (arg0->unk0 + arg0->unk4);
+    sp4 = arg0->unk0 + arg0->unk4;
     arg0->unk4 = sp4;
     return arg0;
 }
